+++ app/constants.py
@@ -1,12 +1,19 @@
-from enum import Enum
-
-DB_NAMING_CONVENTION = {
-    "ix": "%(column_0_label)s_idx",
-    "uq": "%(table_name)s_%(column_0_name)s_key",
-    "ck": "%(table_name)s_%(constraint_name)s_check",
-    "fk": "%(table_name)s_%(column_0_name)s_fkey",
-    "pk": "%(table_name)s_pkey",
-}
+class ResponseField:
+    WORKER = "worker_id"
+    IN = "input_file"
+    JOB_ID = "job_id"
+    IN_STATUS = "input_status"
+    OUT = "analysis_file"
+    OCR = "ocr_file"
+    ANALYSIS = "highlight_file"
+    ANALYSIS_META = "highlight_metadata"
+    TEXT_FILE = "text_file"
+    TEXT = "text"
+    QUALITY = "ocr_quality"
+    STATISTICS = "statistics"
+    TIME = "processing_time"
+    WK_VERSION = "worker_version"
+    KWDS_HASH = "keywords_hash"
 
 
 class Status:
@@ -45,22 +52,3 @@ class JobSpec:
     STATUS = "status"
     ID = "job_id"
     QUALITY = "quality"
-
-
-class Environment(str, Enum):
-    LOCAL = "LOCAL"
-    STAGING = "STAGING"
-    TESTING = "TESTING"
-    PRODUCTION = "PRODUCTION"
-
-    @property
-    def is_debug(self):
-        return self in (self.LOCAL, self.STAGING, self.TESTING)
-
-    @property
-    def is_testing(self):
-        return self == self.TESTING
-
-    @property
-    def is_deployed(self) -> bool:
-        return self in (self.STAGING, self.PRODUCTION)