+++ app/services/ocr_service.py
@@ -1,33 +1,27 @@
 import logging
 import os
 from subprocess import run
+from typing import Tuple
 
 import fitz
 import pikepdf
 from ocrmypdf.__main__ import run as run_ocrmypdf
 from ocrmypdf._exec import tesseract
 
+from app.config import BACK_LANG, LEGAL_LANG, MAX_PAGE_PDF_A, NUM_PROC
 from app.services.text_processing import Cleaner
 
-LEGAL_LANG = "ro_legal"
-BACK_LANG = "ron"
-# number of parallel processes to use for OCR
-NUM_PROC = str(os.environ.get("NUM_PROC", 1))
-
-# maximum number of pages to convert to PDF/A
-# otherwise output type is PDF
-MAX_PAGE_PDF_A = 50
-
-LOGGER = logging.getLogger(__name__)
+logger = logging.getLogger(__name__)
 
 
-def get_language():
+def get_language() -> str:
+    """Get the language to use for OCR."""
     tess_languages = tesseract.get_languages()
     return LEGAL_LANG if LEGAL_LANG in tess_languages else BACK_LANG
 
 
 LANGUAGE = get_language()
-LOGGER.info(f"Using language {LANGUAGE} for OCR")
+logger.info(f"Using language {LANGUAGE} for OCR")
 
 
 WORD_LIST = "nlp/resources/custom-wordlist.txt"
@@ -63,12 +57,13 @@ def is_pdf_valid(input_file: str) -> bool:
         with pikepdf.open(input_file) as _:
             pass
     except pikepdf.PdfError:
-        LOGGER.exception(f"Invalid PDF: {input_file}")
+        logger.exception(f"Invalid PDF: {input_file}")
         return False  # invalid PDF
     return True
 
 
 def is_pdf_encrypted(pdf_file_path: str) -> bool:
+    """Check if a PDF is encrypted."""
     with fitz.Document(pdf_file_path) as doc:
         if doc.needs_pass or doc.metadata is None or doc.is_encrypted:
             return True
@@ -78,33 +73,34 @@ def is_pdf_encrypted(pdf_file_path: str) -> bool:
 
 
 def count_pages(pdf_file_path: str) -> int:
+    """Count the number of pages in a PDF file."""
     with fitz.Document(pdf_file_path) as doc:
         return doc.page_count
 
 
 def remove_encryption(pdf_file_path: str):
-    """Removes encryption by resaving the document
-
-    Args:
-        pdf_file_path (str): pdf file path
-    """
-    LOGGER.info(f"Saving unencrypted document: {pdf_file_path}")
+    """Removes encryption by resaving the document."""
+    logger.info(f"Saving unencrypted document: {pdf_file_path}")
     with pikepdf.open(pdf_file_path, allow_overwriting_input=True) as doc:
         doc.save(pdf_file_path)
 
 
-def make_ocr_command(in_file, pdf_output, pdf_a=True, force_rotate=False):
+def make_ocr_command(
+    in_file: str, pdf_output: str, pdf_a: bool = True, force_rotate: bool = False
+) -> Tuple[str, bool]:
+    """Make the OCR command and check if we have a large number of pages."""
     ocrmypdf_args = [OCRMYPDF, in_file, pdf_output, *CMD_ARGS]
     large_page_count = count_pages(in_file) > MAX_PAGE_PDF_A
     if pdf_a is False or large_page_count:
         ocrmypdf_args = [OCRMYPDF, in_file, pdf_output, *FAIL_SAFE_ARGS, *CMD_ARGS]
     if force_rotate:
         ocrmypdf_args.extend(FORCE_ROTATE_ARGS)
-    LOGGER.debug(" ".join(ocrmypdf_args))
+    logger.debug(" ".join(ocrmypdf_args))
     return ocrmypdf_args, large_page_count
 
 
-def run_ocr_natively(in_file, pdf_output, force_rotate):
+def run_ocr_natively(in_file: str, pdf_output: str, force_rotate: bool):
+    """Run OCR using the ocrmypdf library."""
     ocrmypdf_args, _ = make_ocr_command(
         in_file, pdf_output, pdf_a=False, force_rotate=force_rotate
     )
@@ -113,14 +109,15 @@ def run_ocr_natively(in_file, pdf_output, force_rotate):
         raise Exception("Failed to do OCR.")
 
 
-def call_ocr(in_file, pdf_output, force_rotate):
+def call_ocr(in_file: str, pdf_output: str, force_rotate: bool) -> Tuple[str, str]:
+    """Call OCR using the ocrmypdf subprocess."""
     ocrmypdf_args, _ = make_ocr_command(
         in_file, pdf_output, pdf_a=False, force_rotate=force_rotate
     )
     proc = run(ocrmypdf_args, capture_output=True, encoding="utf-8")
     if proc.returncode != 0:
-        LOGGER.error(proc.stdout)
-        LOGGER.error(proc.stderr)
+        logger.error(proc.stdout)
+        logger.error(proc.stderr)
         # sometimes the PDF is not entirely valid according to the spec
         # but the generated PDF can still be used and rendered
         if not is_pdf_valid(pdf_output):
@@ -128,7 +125,8 @@ def call_ocr(in_file, pdf_output, force_rotate):
     return proc.stdout, proc.stderr
 
 
-def get_ocrized_text(pdf_file):
+def get_ocrized_text(pdf_file: str) -> str:
+    """Get the OCRized text from a PDF file."""
     text = ""
     with fitz.open(pdf_file) as pdf_f:
         for page in pdf_f.pages():
@@ -136,7 +134,8 @@ def get_ocrized_text(pdf_file):
     return text
 
 
-def get_ocrized_text_from_blocks(pdf_file):
+def get_ocrized_text_from_blocks(pdf_file: str) -> str:
+    """Get the OCRized text from a PDF file in a clean format using blocks."""
     text = ""
     with fitz.open(pdf_file) as pdf_f:
         for page in pdf_f.pages():
@@ -146,12 +145,7 @@ def get_ocrized_text_from_blocks(pdf_file):
     return text
 
 
-def extract_ocrized_text(pdf_file, txt_output_file):
-    text = get_ocrized_text(pdf_file)
-    dump_text(text, txt_output_file)
-    return text
-
-
-def dump_text(text, txt_output_file):
+def dump_text(text: str, txt_output_file: str):
+    """Dump text to a file."""
     with open(txt_output_file, "w", encoding="utf-8") as txt_f:
         txt_f.write(text)